fixing contract graphql

+++ features/merchant/product_detail_common/src/main/java/com/tokopedia/product/detail/common/data/model/Cashback.kt
@@ -31,5 +31,5 @@ data class Wholesale(
 
         @SerializedName("price")
         @Expose
-        val price: Int = 0
+        val price: Float = 0f
 )
